# docs/api-docs.yaml
openapi: 3.0.3
info:
  title: API de Carrito y Órdenes
  version: 1.0.0
  description: Microservicio NestJS para gestionar carritos de compra y órdenes en PulgaShop
  contact:
    name: Equipo PulgaShop
servers:
  - url: http://localhost:4000/api/v1
    description: Local
  - url: https://pulga-shop-carrito-ordenes-api.onrender.com/api/v1
    description: Production
tags:
  - name: Carrito
    description: Endpoints de carrito
  - name: Órdenes
    description: Endpoints de órdenes
paths:
  /carrito/crearCarrito:
    post:
      tags: [Carrito]
      summary: Crear carrito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearCarritoRequest'
      responses:
        '201':
          description: Carrito creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarritoDetallado'
  /carrito/productos:
    get:
      tags: [Carrito]
      summary: Listar productos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: take
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Productos obtenidos
  /carrito/obtenerCarritos:
    get:
      tags: [Carrito]
      summary: Obtener todos los carritos
      responses:
        '200':
          description: Lista de carritos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CarritoDetallado'
  /carrito/obtenerCarro/{id}:
    get:
      tags: [Carrito]
      summary: Obtener carrito por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Carrito encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarritoDetallado'
  /carrito/agregarProductos:
    post:
      tags: [Carrito]
      summary: Agregar producto al carrito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgregarProductoRequest'
      responses:
        '201':
          description: Producto agregado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarritoDetallado'
  /carrito/eliminarProducto:
    delete:
      tags: [Carrito]
      summary: Eliminar producto del carrito
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EliminarProductoRequest'
      responses:
        '200':
          description: Producto eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarritoDetallado'
  /carrito/eliminarCarrito:
    delete:
      tags: [Carrito]
      summary: Eliminar carrito por compradorId
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EliminarCarritoRequest'
      responses:
        '200':
          description: Carrito eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarritoDetallado'
  /carrito/eliminarCarrito/{id}:
    delete:
      tags: [Carrito]
      summary: Eliminar carrito por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Carrito eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarritoDetallado'
  /ordenes/checkout/{carritoId}:
    post:
      tags: [Órdenes]
      summary: Crear orden (checkout)
      parameters:
        - name: carritoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Orden creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenCheckoutResponse'
  /ordenes/estado-pago/{carritoId}:
    get:
      tags: [Órdenes]
      summary: Obtener estado de pago
      parameters:
        - name: carritoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estado de pago
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadoPagoResponse'
  /ordenes/historial/{compradorId}:
    get:
      tags: [Órdenes]
      summary: Obtener historial de compras
      parameters:
        - name: compradorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Historial obtenido
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrdenConItems'
  /ordenes/estadisticas/producto/{productoId}:
    get:
      tags: [Órdenes]
      summary: Obtener estadísticas de producto
      parameters:
        - name: productoId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estadísticas obtenidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EstadisticasProductoResponse'
  /ordenes/actualizarCheckout/{ordenId}:
    put:
      tags: [Órdenes]
      summary: Actualizar orden
      parameters:
        - name: ordenId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarCheckoutRequest'
      responses:
        '200':
          description: Orden actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrdenCheckoutResponse'
components:
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
    CrearCarritoRequest:
      type: object
      properties:
        compradorId:
          type: string
      required:
        - compradorId
    EliminarCarritoRequest:
      type: object
      properties:
        compradorId:
          type: string
      required:
        - compradorId
    AgregarProductoRequest:
      type: object
      properties:
        compradorId:
          type: string
        productoId:
          type: integer
        cantidad:
          type: integer
      required:
        - compradorId
        - cantidad
    EliminarProductoRequest:
      type: object
      properties:
        compradorId:
          type: string
        productoId:
          type: integer
        cantidad:
          type: integer
      required:
        - compradorId
        - cantidad
    CarritoDetallado:
      type: object
      properties:
        id:
          type: integer
        compradorId:
          type: string
        total:
          type: number
        items:
          type: array
          items:
            type: object
    OrdenCheckoutResponse:
      type: object
      properties:
        orden:
          type: object
        items:
          type: array
        total:
          type: number
    EstadoPagoResponse:
      type: object
      properties:
        ordenId:
          type: integer
        estadoPago:
          type: string
        total:
          type: number
    OrdenConItems:
      type: object
      properties:
        orden:
          type: object
        items:
          type: array
        total:
          type: number
    ActualizarCheckoutRequest:
      type: object
      properties:
        items:
          type: array
      required:
        - items
    EstadisticasProductoResponse:
      type: object
      properties:
        vecesComprado:
          type: integer
        vecesCancelado:
          type: integer